requirements.txt
streamlit
Pillow
opencv-python-headless
speechrecognition
